+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/api/WebResources.scala
@@ -3,9 +3,12 @@ package pl.touk.nussknacker.ui.api
 import akka.http.scaladsl.model.headers.{CacheDirectives, `Cache-Control`}
 import akka.http.scaladsl.server.{Directives, Route}
 import com.typesafe.scalalogging.LazyLogging
-import org.apache.commons.io.{FileUtils, IOUtils}
+import org.apache.commons.io.FileUtils
+import pl.touk.nussknacker.engine.util.ResourceLoader
 
+import java.nio.charset.StandardCharsets
 import java.nio.file.Files
+import scala.util.Try
 
 class WebResources(publicPath: String) extends Directives with LazyLogging {
 
@@ -13,14 +16,15 @@ class WebResources(publicPath: String) extends Directives with LazyLogging {
   private lazy val mainContentFile = {
     val tempMainContentFile = Files.createTempFile("nussknacker", "main.html").toFile
     tempMainContentFile.deleteOnExit()
-    val data = getClass.getClassLoader.getResourceAsStream("web/static/main.html")
-    val content = Option(data).map(IOUtils.toString).getOrElse {
+    val data = Try(ResourceLoader.load("/web/static/main.html"))
+
+    val content = data.toOption.getOrElse {
       logger.error("Failed to find web/static/main.html - probably frontend resources are not packaged in jar. Frontend won't work properly!")
       ""
     }
     val withPublicPathSubstituted = content.replace("__publicPath__", publicPath)
 
-    FileUtils.writeStringToFile(tempMainContentFile, withPublicPathSubstituted)
+    FileUtils.writeStringToFile(tempMainContentFile, withPublicPathSubstituted, StandardCharsets.UTF_8)
     tempMainContentFile
   }